chore: Rename table to eventAttendanceTable for clarity
